EnumModel, EnumUtils 제거

+++ src/main/java/io/hurem/domain/employees/EmployeeStatusRepository.java
@@ -2,4 +2,4 @@ package io.hurem.domain.employees;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmployeeStatusRepository extends JpaRepository<EmployeeStatus, Integer> { }
+public interface EmployeeStatusRepository extends JpaRepository<EmployeeStatus, Long> { }
